requirements.txt
aiohttp~=3.7.4.post0
PyAudio~=0.2.11
pycryptodome~=3.10.1